# demo.Web.Host/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "参考ABP 框架搭建",
        "title": "GoLangWebFrameWork",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "1362",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "paths": {
        "/User": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户相关"
                ],
                "summary": "新增用户信息",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UserAddInputDto"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserLogin 登录相关"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Success\": true,\"Token\":\"5545\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"who are you\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 获取用户列表"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "dto.UserAddInputDto": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "电话",
                    "type": "string"
                },
                "sex": {
                    "description": "性别",
                    "type": "integer"
                }
            }
        },
        "dto.UserLoginInputDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}